// utils/enum.ts
export enum ProductCategory {
    ALL = "All",
    ELECTRONICS = "electronics",
    JEWELERY = "jewelery",
    MENS_CLOTHING = "men's clothing",
    WOMENS_CLOTHING = "women's clothing",
}

export enum ProductSortOption {
    PRICE_LOW_TO_HIGH = "Price - Low to High",
    PRICE_HIGH_TO_LOW = "Price - High to Low",
    HIGHEST_RATED = "Highest Rated",
    NAME_A_TO_Z = "Name - A to Z",
    NAME_Z_TO_A = "Name - Z to A",
}